Ensures all tests except the synth tests pass in Node.js.
