+++ tests/unit/all-tests.html
@@ -24,6 +24,7 @@
         "./html/core-utilities-tests.html",
         "./html/dynamics-ugen-tests.html",
         "./html/envelope-ugen-tests.html",
+        "./html/envgen-ugen-tests.html",
         "./html/filter-ugen-tests.html",
         "./html/gate-ugen-tests.html",
         "./html/ioc-integration-tests.html",
