+++ tests/unit/nodejs/all-node-tests.js
@@ -34,27 +34,28 @@ var testIncludes = [
     "../js/core-tests.js",
     "../js/core-utilities-tests.js",
     "../js/dynamics-ugen-tests.js",
-    "../js/envelope-ugen-tests.js", // TODO: These tests poison the subsequent tests.
-    // "../js/filter-ugen-tests.js",
-    // "../js/gate-ugen-tests.js",
+    "../js/envelope-ugen-tests.js",
+    "../js/envgen-ugen-tests.js",
+    "../js/filter-ugen-tests.js",
+    "../js/gate-ugen-tests.js",
     "../js/ioc-integration-tests.js",
-    // "../js/listening-ugen-tests.js",
-    // "../js/math-ugen-tests.js",
+    "../js/listening-ugen-tests.js",
+    "../js/math-ugen-tests.js",
     "../js/midi-tests.js",
-    // "../js/model-synth-tests.js",
-    // "../js/multichannel-ugen-tests.js",
+    "../js/model-synth-tests.js",
+    "../js/multichannel-ugen-tests.js",
     "../js/nodelist-tests.js",
-    // "../js/osc-ugen-tests.js",
+    "../js/osc-ugen-tests.js",
     "../js/parser-tests.js",
-    // "../js/phasor-ugen-tests.js",
+    "../js/phasor-ugen-tests.js",
     "../js/playbuffer-ugen-tests.js",
-    // "../js/random-ugen-tests.js",
+    "../js/random-ugen-tests.js",
     "../js/scheduler-tests.js",
     "../js/scheduler-timing-tests.js",
-    // "../js/scheduling-ugen-tests.js",
+    "../js/scheduling-ugen-tests.js",
     // "../js/synth-tests.js",
-    // "../js/trigger-ugen-tests.js",
-    // "../js/ugen-tests.js"
+    "../js/trigger-ugen-tests.js",
+    "../js/ugen-tests.js",
     "../js/writebuffer-ugen-tests.js"
 ];
